The current std::hash specialization for winrt::hstring simply calls the same _Hash_seq function that the std::basic_string specialization uses.
There are two problems with this. Firstly, we should avoid relying on an implementation detail of the VC libraries that may change in future. Secondly, this makes it harder to support C++/WinRT on other platforms.

Stephan T. Lavavej suggested that we provide our own implementation. I have thus copied VC’s FNV-1a implementation and updated it to match the C++/WinRT coding style.

I have also removed the dependence on std::unary_function since that is no longer available in VC Update 3 with /std:c++latest.

The unit tests merely confirm that the results match that of std::wstring. It is possible that VC’s implementation may change in future, but I’d rather our unit tests provide an early warning of that change so that we may consider making a similar change if appropriate.