Move samples to a folder name samples.
changes to all samples so that they can build after the move of all the samples to a new directory

More changes to the sample projects to more properly point at where the headers are located.

changes to all samples so that they can build after the move of all the samples to a new directory

ammend previous merge

Move samples to a folder name samples.

Move samples to a folder name samples.

changes to all samples so that they can build after the move of all the samples to a new directory

Move samples to a folder name samples.

Changes to address bad merges

Move samples to a folder name samples.