Merged PR 402767: Support for generic required interfaces
I have simpler version of this in the works but it will likely take another week or two to fully bake. In the mean time, this handles generic interfaces when they are present as required interfaces so that types are properly instantiated when aggregating generic and non-generic types and the necessary using directives are employed to disambiguate and fold in inherited base classes.