Merged PR 433436: Merge kitchen_sample to dev
This is a port of the [Windows UWP SDK Kitchen sample](https://github.com/Microsoft/Windows-universal-samples/tree/master/Samples/WRLInProcessWinRTComponent) to C++/WinRT. This provides us a simple WinRT component that can be implemented in pure standard (aka xplat) C++.

Note, the dev branch cppwinrt compiler component support is currently broken. The checked in code was generated w/ an earlier version of the compiler. The kitchen component runs cppwinrt as part of the build process, but that step is currently commented out until the compiler is updated.