Merged PR 444746: Integrate std::wstring_view and concatenation to hstring
This incorporates std::wstring_view into winrt, allowing us to convert between hstring and wstring_view, as well as compare them against each other. Doing this also simplified the work to make hstring concatenation function, so I threw that on the pile.

One complication was that wstring_view supports implicit conversions from string literals, basic_string, const wchar_t*, etc, so I ran into rampant overload lookup ambiguities at first. Solving this required making hstring's constructors explicit. I'm fine with this design, as constructing an hstring from a string literal incurs an expensive call to WindowsCreateString. Also, this is the pattern followed with most other abstractions of this kind - allow implicit conversions between cheap helper classes, but force users to opt-in to an expensive conversion. In a nutshell, this now requires us to explicitly construct an hstring when one is needed, such as when returning an hstring from GetRuntimeClassName(). That sort of fixup comprises most of the test churn.

Another side effect was that I had to start peeling apart hstring and hstring_view. With hstring_view greedily participating in implicit conversions (by design), and implicitly converting to hstring, this was another source of ambiguous overload lookup, as the compiler couldn't decide whether to convert a string literal into a wstring_view to pass to hstring, or to construct an hstring_view to convert to hstring. This separation was necessary anyway, and moves hstring_view further down the road of becoming strictly an input parameter wrapper, and not being consumed by developers at all, which is our eventual goal.

Related work items: #11722121, #11721465