Merged PR 490732: WINRT_DISABLE_FACTORY_CACHE
We have not yet fixed the relationship between CRT startup and initialization and destruction of COM statics, so we need a way for developers to opt out of using factory caching for correctness. Defining WINRT_DISABLE_FACTORY_CACHE will disable the caching and thus make it safe for use in DLLs.