Merged PR 490313: Replacement generator that supports exceptions
The std::experimental generator does not correctly handle exceptions. This will be fixed in 2017.6 but in the mean time it causes problems for us when we need to propagate error information back to the caller from within the cppwinrt compiler. This fixes that. In the future we can switch back to the VC generator if appropriate.