Merged PR 495173: Input parameters
This PR is mostly about restoring functionality that never made it from the old compiler into the new compiler. It deals with handling input parameters of string and collection types and the usability affordances for binding to standard C++ types. It also includes some improvements not previously available to avoid usability problems and improve the usability in general. Highlights:

•	Moved the hand-rolled base.h (pre-amalgamation) that both the base project and the compiler now share into a single base.h header so that it’s no longer duplicated. This just eases maintenance for us internally.

•	Updated the code writer’s write_param_type to use the new param types as appropriate.

•	Added a separate write_component_param_type function for use by implementations where you simply want to use the actual projection types.

The param types are as follows. Note that these are only used on the consumption surface and only applies to input parameters. We also distinguish between synchronous method calls and what *might* be asynchronous. An asynchronous method is one that ends with “Async” or is implementing [propput]. For these we err on the side of usability to opt not to offer the no-copy semantics by default.

hstring – replaced with “param::hstring const&”. This is basically what hstring_view did before with a few notable changes. It’s always const& to avoid copies (however trivial) when an interface/class member calls an underlying factory member. Since the param::hstring is never actually used by developers, it lacks all of the conveniences of hstring itself and we don’t have to provide all of the operators either.

IIterable<T> - replaced with “param::iterable<T> const&” for sync methods and  “param::async_iterable<T> const&” for async methods. The latter doesn’t provide no-copy constructors.

The same applies to IVectorView and IMapView. For IVector and IMap (rather rare as input parameters) there is no async variant since they are by definition mutable and having a no-copy version would typically be unexpected.