Merged PR 556599: Better handling of missing out params in composable factories
If I construct a composable type, the projection goes through that type's composable factory interface, and supplies an "outer" input parameter, and an "inner" output parameter to handle the COM aggregation rules. In C++/CX, if a type hasn't been subclassed, it neglects to supply the "inner" output parameter, leading to a crash if the interface's implementation doesn't check the validity of the param.

C++/CX is most likely wrong here, but the reality is that even if it does get fixed, we'll still need to interop with code written from an older, unfixed compiler. So, unfortunately, we need to bite the bullet and handle this situation gracefully.

Related work items: #12334358