Merged PR 638156: Component tweaks including UTF-8 support
This update includes a few small improvements to the component authoring experience including support for UTF-8 identifiers. The cppwinrt compiler now emits the UTF-8 preamble (BOM), which appears to be the only reliable way to get the Visual C++ compiler to recognize the character set.

I've also added a pragma warning suppress to the xaml shim to avoid the warning generated on Update 2 for the new Update 3 _has_include identifier.

Finally, I cleaned up and updated the Authoring sample to provide an example of UTF-8 support.