Merged PR 646072: Validate interface input parameters
This update provides a minor addition to the validation done when producing interface implementations. Interface input parameters are now checked for nullptr. If so, an hresult_invalid_argument is thrown with the name of the parameter. This ensures that any language projection will get both the "invalid argument" message as well as the name of the offending parameter.

The unit tests were updated to include a specific test for this. The event_consume test was also updated since it was passing a nullptr as part of its test code. :)