Merged PR 655835: Natural conversion between implementation and class types
The implements class template ensures that you can naturally convert from an implementation to any of the interfaces that the implementation supports. The runtime class type is however not one of those types and this makes it’s a little awkward for component authors. This update solves that problem by extending this to support class types for implementations and updates the make function to return the class type rather than merely the default interface. Both are major usability wins for component authors, particularly in scenarios where events are implemented and the “sender” must be provided as a class type and custom event args are constructed within a component. The authoring sample has also been updated with a slightly more interesting IDL scenario to test more of the edge cases in the cppwinrt compiler including the scenarios addressed in this update.