Merged PR 664163: Support for structs by ref
WinRT supports passing struct parameters “by reference” even though this feature has never used by the Windows SDK. By reference means that the struct is not copied on the ABI, as is the default. This update adds support for such parameters as they have been found to exist in some internal metadata used within Windows.

The change is relatively minor. We detect the optional modifier and simply add a const& to the parameter for the ABI’s signature. This prevents the C++ compiler from injecting the equivalent of a call to the struct copy assignment operator. Avoiding this copy is why this feature was first introduced.

The optional modifier does not mean that the parameter is optional, merely that CLI consumers may ignore the modifier if necessary. WinRT does not support any other modifiers at present. We may need to harden this implementation further to support future modifiers, but I’ll leave that until we refactor the signature parsing code which needs some work.

The Authoring sample has been updated to illustrate its use from C#. Notably, C# does not allow a null value to be passed for such parameters so it does honor the same “a value is expected” semantics that C++ defines for const& parameters.