Merged PR 669971: Fixed all includes to be relative to winrt folder
The relative/component projections must be able to refer back to the platform projection without actually being copied into the same folder structure. This update fixes a few places where the headers were being generated with relative paths assuming that they would all be located in the same folder. With this change, we can now generate projections for libraries like win2d and have them be distributed separately. Inclusion is a matter of updating the VC project settings to include the additional paths.