Merged PR 673608: Complex component sample and bug fixes
I’ve removed the edge case testing from the Authoring sample to keep it nice and simple. Instead, I’ve added a Complex.Authoring sample that’s dedicated to validating various scenarios unique to component authoring. This includes:

* Simple class named “Simple”
* Language class validating UTF-8 support
* StructByRef class validating support for struct-by-ref parameters
* Nested class also named “Simple” validating support for classes in multiple namespaces
* Static class validating support for static class generation

These scenarios, particularly the last two, uncovered a number of bugs in our handling of paths and type names that are also addressed in this update. More work needs to be done but this update sets the stage for further component authoring validation and usability improvements.