Merged PR 679445: Proper filtering for component authoring
The -filter option was originally conceived as a way to project a subset of types for a given implementation. You might want to implement just the Uri class so you might include the following options:

cppwinrt.exe -component -filter Windows.Foundation.Uri -ref local

It would generate the scaffolding necessary for you to implement the Uri class itself without having to implement the rest of the classes in the winmd that defines the Uri class. Unfortunately, the implementation conflated filtering with reference types (the -ref keyword) and so this didn’t work correctly. I’ve now corrected this and renamed some of the internal implementation details to more accurately reflect their meaning. For example, get_unfiltered_types is now called get_projected_types and type.filtered is now type.is_reference in line with database.is_reference.

This update also includes a number of additional fixes for component authoring along with some more “complex authoring” samples to illustrate and validate this. This includes support for implementing runtime classes in different namespaces and winmds that include multiple namespaces as well as validating support for factory name mangling.