Merged PR 714156: Simpler com_ptr accessors and Direct2D sample
I've had a lot of requests lately for a Direct2D sample. Given than C++/WinRT is the "systems programming language" for the Windows Runtime, it makes sense that it would be ideally suited to this kind of high performance app. I made some small improvements to the com_ptr accessor functions and wrote a complete Direct2D sample app, which provides a good example of com_ptr use in a realistic setting.