Merged PR 759757: Namespace paritioning for module support
This is a benign change that simply refactors namespace declarations so that we can export the winrt namespace as a module without exporting the nested winrt::impl namespace. The export keyword includes all nested namespaces in the module and does not provide an exclusion mechanism. For this reason, we can simply restructure our namespace declarations to avoid exporting the impl namespace.

Anyway, this PR is just to contain that refactoring of namespaces so that subsequent updates can focus on targeted support for modules. Finally, I'd like to switch our unit test project to using modules exclusively. That's coming up.