Merged PR 765518: Eliminate redundant instructions in param::hstring constructors
The DEP controls team pointed out some bloat when using the new projection. This is an easy one to fix - we're always zero-filling the HSTRING and HSTRING_HEADER fields right before populating them with the results of WindowsCreateStringReference.

This added "only" 16 bytes of instructions to the method, but because this is a small function that's almost always inlined, this 16 bytes was showing up repeatedly all over the place, so that it was adding nearly 30kB of useless bloat to a 3MB dll (1%), accounting for nearly 10% of the code size regression.