Merged PR 769852: Strike down guid_v-related executable bloat
Change guid_v to a constexpr method (similar to __uuidof()) to eliminate executable bloat from const data duplication.

We can't use __declspec(selectany) because it only works on items with external linkage. But, if it's simply a constexpr method that returns the GUID, those functions will get folded. And if the method returns a GUID const&, we still get the folding without the overhead of an additional copy (I disassembled the updated DEP controls dll to verify).