Merged PR 771415: Added declspec(selectany) extern string data to ID C++/WinRT library version in use.
Added declspec(selectany) extern string data to ID C++/WinRT library version in use.  Pragma detect_mismatch also discussed, as it could do double-duty of enforcing a consistent library, but this artifact disappears at link time, so can't be used in store ingestion.  Also, while the WINRT_version symbol exists in the PDB, this is optional for store submissions.  So the prefix is necessary in the string data (and is subject to false positives, though unlikely).  The comment in base_version.h per Ben's request, and suggests that the data could be removed (I originally had an opt-out #define, but Ben preferred the stronger approach).

Related work items: #13166793