Merged PR 864771: noexcept
This update applies the noexcept specifier to properties as well as supporting internals to increase the likelihood that more implementations will be considered noexcept by the C++ optimizer and thereby reduce the exception handling overhead within components.

This update also includes a small change to the IReference<T> implementation to support XAML and some other minor conformance changes.