Merged PR 877795: Added missing IPropertyValue required interface and optimized IReference
Added missing IPropertyValue required interface and optimized IReference implemenation for custom types. We previously did a lot of work to get the IReference<T> defined outside of base.h so that we can correctly require IPropertyValue, but then we never looped back and actually added the required interface. I've corrected that here.

I have also optimized the delegating constructors in IReference so that instantiations for custom types don't introduce a redundant call to QueryInterface. The QI for IReference is sadly required by PropertyValue statics but not for the custom IReference implementation provided directly by C++/WinRT.