Merged PR 905582: Workaround for VC bug in com_array::clear
There's a bug in VC code gen following an if constexpr where it injects a temporary that never gets destroyed. The result is that the clear method copies and then destroys and if that's a COM object it basically calls AddRef and then Release. One solution could be to use auto&&, which basically avoids the problem but C++17 introduces std::destroy and that ends up being much simpler. It does all of the heavy lifting and optimizations for us.