Merged PR 919361: Generate impl::require and impl::base for component authored classes.
When implementing a component that composes a base type, (XAML controls), authors didn't have use of the impl::require and impl::base helpers that provide a straightforward way of converting to base types and calling methods on those base types' interfaces.

This change incorporates those metaclasses into the component authoring experience. That logic was getting more and more untenable, so I cleaned some house in that area to put this in.

Related work items: #13839990, #13882567