Merged PR 941192: Compact struct ABI to reduce dependencies
This addresses the inability to support internal metadata due to the complex_structs.h compilation errors. The solution is to remove all ABI type information for structs and replace them with generic struct_of<T> value types that are trivially copyable and thus give the correct ABI semantics without imposing the interdependencies caused by having actual ABI structs. This is similar to the way we treat all class and interface parameters as IUnknown interface pointers on the ABI.