Merged PR 968769: De-inline calls to Release in dtors of com_ptr and IUnknown.
De-inline calls to Release in dtors of com_ptr and IUnknown. We generate a lot of temporaries that get moved from. This results in a lot of dtor calls that only need the initial null check, but the whole dtor gets inlined at all those call sites.

Separating the null check and the rest of the dtor into separate methods allows us to inline the small null check, and put the rest of the cleanup into a non-inlined method, which yields significant code size savings.