Merged PR 1021222: Folded the implementation of hstring.
This is a simple housekeeping update. Originally, C++/WinRT had two string classes - one for ref-counted hstrings and one for fast-pass hstrings. They provided convertibility in both directions, which meant they needed to be defined up front and member definitions came later. We've since moved to a single (public) hstring class for simplicity, but I never got around to folding the member definitions back into the hstring class for simplicity and consistency. That's what's happening here.

I've also removed the implicit constructor for std::wstring - this was a bug. It was meant to be removed when we added the std::wstring_view constructor. Implicit conversions make it way too easy to deep copy strings and containers of strings accidentally. This also follows the pattern set by the C++ standard library.