Merged PR 1044537: Correct logic controlling the component boundary
This fixes up the logic that controls COM aggregation logic at the component boundary.

Specifically, a type is only considered to be "outside this component" if it's in a "reference" winmd, or if there are filters present and the type is outside the filters.

Another way of putting it is that a type is considered to be "inside this component" if it's not in a "reference" winmd and either (1) no filters are present or (2) the type is in at least one filter.

Note that we still need to refine our filtering logic - our substring matching may be overly aggressive. For example, if I have a WidgetBase in one component and a Widget in another, I currently have no way to specify that I want to

Related work items: #14288831