Merged PR 1066909: Prevent put_abi misuse
This update fixes the issue where put_abi returned an ::IUnknown** that could be misinterpreted by libraries that use __uuidof to determine the type of a parameter and thus cause unexpected runtime failures. Instead, put_abi now simply returns void** and the developer is greeted with a compiler error instead. The prescribed ABI interop code continues to work as before.

I've also fixed Clang support that was broken with #pragma detect_mismatch which doesn't behave quite the same with Clang.

I also removed some dead code for the internal handle class and simplified the ABI for IUnknown based types in general.