Merged PR 1100949: Testing via desktop bridge
This is a new project configuration for testing C++/WinRT. The existing unit tests are limited in that they cannot use custom components since they run in a console app. This means we need to rely on a variety of OS components to test various aspects of the projection, but that makes the tests rather contrived and forces a lot of OS headers to be included which slows down the build time. We also rely on building our sample projects to validate various things that the unit tests cannot validate. This is very error prone.

This new project can run both traditional unit tests (like hstring operators) and WinRT component testing since it runs from a packaged app. I'd like to eventually move all our unit tests over to this project.