Merged PR 1132611: Errors
I've added comprehensive testing for error propagation to the bridge tests. I've also removed the RPC_E_DISCONNECTED mapping since we don't use it and there's no reason a developer should raise or even catch this. I was recently reminded that the HRESULT itself is only for diagnostics and the WinRT error model is centered around RoOriginate and the message it propagates rather than the HRESULT so there's no reason to map every conceivable HRESULT, nor is that practical. I've also hoisted to_hresult out of the impl namespace as it is essential for interop where the developer has their own ABI and needs some way to turn exceptions into HRESULTs in a consistent manner.