Merged PR 1142393: Safe coroutines by default
A common pitfall with coroutines is to neglect to declare parameters by value such that the coroutine will hold a value rather than a reference to the input. This isn't always a bug but it often is. This update simply ensures that the starter code we generate for component authors does the safe thing by default. For example, here's a non-coroutine that's efficient by default:

```C++
void Test(hstring const& a,
          int32_t b,
          Windows::Foundation::IAsyncInfo const& c,
          Windows::Foundation::Rect const& d);
```

And here's what we generate for coroutines. Notice that all the &'s are gone:

```C++
Windows::Foundation::IAsyncAction TestAsync(hstring const a,
                                            int32_t b,
                                            Windows::Foundation::IAsyncInfo const c,
                                            Windows::Foundation::Rect const d);
```

Naturally, a developer can override the defaults but at least these defaults are a good starting point to avoid bugs.