Merged PR 1155235: Internal metadata does not compile: runtime class has multiple interfaces but none are default
This update adds support for runtime classes that have multiple interfaces but none are marked the default. We now infer the default interface based on the first interface returned by EnumInterfaceImpls as being the first listed in IDL and thus assumed to be the default if there is no explicit default interface.

Related work items: #14857816