Merged PR 1164719: Support for static lifetime of activation factories
The “StaticLifetimeStore” is used to ensure that factories that hold state remain alive for the duration of the MTA so that they can correctly implement things like events even when the application is not explicitly keeping the factory alive. Without some form of static lifetime, any state held by the factory would be lost if not backed by some other persistent storage.

This update adds the static_lifetime method that a factory may call during construction to ensure that it remains alive until the MTA is torn down. This is achieved by using the CoreApplication activation factory, whose lifetime is itself tied to the MTA, and works perfectly well in both packaged and desktop apps. The CoreApplication factory maintains a map of <hstring, IInspectable> pairs that objects may enlist in to ensure that the CoreApplication factory keeps them alive simply by virtue of holding a reference. The static_lifetime method simply enlists the object itself in the collection using the object’s runtime class name as the key. The implementation (in base.h) avoids a dependency on the Windows.ApplicationModel.Core.h header by calling RoGetActivationFactory directly. The ICoreApplicationPrivate interface is undocumented and contains unrelated functionality that I don’t wish to make public. I have thus called it IStaticLifetime and placed it in the impl namespace so that we can extend/rename it at a later stage if needed.

Consider this simple factory that does not have static lifetime:

```C++
struct Factory : implements<Factory, IActivationFactory>
{
    Factory()
    {
        puts("Factory");
    }

    ~Factory()
    {
        puts("~Factory");
    }

    IInspectable ActivateInstance()
    {
        throw hresult_not_implemented();
    }
};
```

Now consider this application:

```C++
int main()
{
    puts("init_apartment");
    init_apartment();

    {
        auto f = make<Factory>();
    }

    puts("uninit_apartment");
    uninit_apartment();
}
```

The init_apartment call will create the MTA and the uninit_apartment call will tear down the MTA. Of course, the lifetime of the factory is neatly bound by the inner scope, producing the following output:

```
init_apartment
Factory
~Factory
uninit_apartment
```

Now let’s update the factory to have static lifetime:

```C++
struct Factory : implements<Factory, IActivationFactory>
{
    Factory()
    {
        puts("Factory");
        static_lifetime(); // <-- here!
    }
.
.
.
```

Without changing the application, the lifetime of the factory is extended to that of the MTA:

```
init_apartment
Factory
uninit_apartment
~Factory
```

I have also updated the BridgeTests to properly handle lifetime for static events.

Related work items: #14817230