Merged PR 1167569: Faster events by default
WinRT events are slow. While chatting about this, Ben had a great idea to improve event performance in the short term.

The basic idea is to only ever store agile delegates. If a non-agile delegate arrives, we simply wrap it inside an agile delegate before storing it in the event array. That way, the cost of unwrapping the delegate is paid only by those with non-agile delegates rather than penalizing those who have agile delegates to begin with.

I took a quick look and sure enough it almost doubles the invocation speed given agile delegates. This also performs at least as well as the non_agile_event implementation, which allows us to get rid of a confusing part of the implementation story: which event class template should I use? Now we only have one with the obvious name `event<Delegate>` and the non-obvious fallback for those rare single-threaded APIs namely `single_threaded_event<Delegate>`. This is a name change, which may affect a few early adopters but I believe the performance win is compelling and the reduced confusion around naming and choices is well worth it at this early stage.