Merged PR 1274478: resume_foreground
`resume_foreground` is a complement to C++/WinRT's `resume_background` coroutine helper. The base library currently provides `resume_background` to switch a coroutine to the thread pool, but there's only an indirect way to switch to a UI thread via the `apartment_context` helper.

Keep in mind that `apartment_context` is still preferred if the coroutine begins life on the UI thread, but in the event that the coroutine does not begin on the UI thread, it must use the `CoreDispatcher` to find its way there. That's what `resume_foreground` does.

Also, note that if the coroutine begins life on the UI thread and you never use `resume_background` then there's never any need to use either `apartment_context` or `resume_foreground` since the coroutine will always resume on the UI thread after every suspension point.

The challenge is that the implementation relies on both `Windows.UI.Core` and `Windows.ApplicationModel.Core` as illustrated here:

https://gist.github.com/kennykerr/341be078f87e871edc6e8c2ad5608e92

Thus it cannot live beside `resume_background` in the `base.h` header, but must trail both of these headers. I picked `Windows.ApplicationModel.Core` as it already has a dependency on `Windows.UI.Core` but the inverse is not true. Once we support C++ modules, adding these sorts of helpers will be much easier as we won't have the namespace split headers to contend with.

Related work items: #15184250