Merged PR 1299101: Fix async component signature generation
The has_reference_param function was originally conceived to support the code gen for optimizing input parameter bindings - that's all. I later used it "conveniently" to ensure that the component authoring stubs we produce don't use references for async methods, since that can lead to hard-to-debug errors in coroutines. The trouble is that the two purposes weren't identical and the has_reference_param helper was horribly misleading. The parameter optimization covers property putters but the component author should continue to use const& parameters for property implementations.

Related work items: #15349576