Merged PR 1308139: Make it easier to use C++/WinRT to manage call context within services
Coroutines can make it tricky to manage call context and impersonation since you can easily transition to a different thread and thus lose the ability to impersonate the client. This update adds the `access_token` class and updates the `resume_background` coroutine helper by adding a simple form of customization of the background execution context.

The `access_token` class models a Windows access token (e.g. security context) representing the identity and privileges of a user account. It is just a simple extension of the C++/WinRT `handle` class, so you can use it like a regular `handle` to manage ownership and lifetime of an exclusive handle type.